LayoutTests/accessibility/aria-disabled-expected.txt


This tests that the aria-disabled attribute works. The text field should not be enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS succeeded is false
PASS succeeded is true
PASS successfullyParsed is true

TEST COMPLETE
